add a parent mushroom class
